# php/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-BookingHealthcare API - Swagger",
        "description": "E-BookingHealthcare API with Laravel",
        "contact": {
            "name": "Nguyen Chinh",
            "url": "https://github.com/ngchinhdev/",
            "email": "chinhnguyennn24@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/services": {
            "get": {
                "tags": [
                    "Service Routes"
                ],
                "summary": "Get all Service",
                "operationId": "63d1ccd0611b736ba3f45d82a1c1bd7d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/services/{id}": {
            "get": {
                "tags": [
                    "Service Routes"
                ],
                "summary": "Get one Service",
                "operationId": "b410ce76beeedf466d89feeb33cb9938",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/services/add": {
            "post": {
                "tags": [
                    "Service Routes"
                ],
                "summary": "Add a Service",
                "operationId": "3708ab2fa6f9966f290f8bc99614c8fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "specialtyID": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Name of the new Service"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 32443242
                                    },
                                    "shortDescription": {
                                        "type": "string",
                                        "example": "content of the new Service"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "author of the new Service"
                                    },
                                    "detail": {
                                        "type": "string",
                                        "example": "detail"
                                    },
                                    "discountPrice": {
                                        "type": "number",
                                        "example": 312412
                                    },
                                    "duration": {
                                        "type": "number",
                                        "example": 2342
                                    },
                                    "isHidden": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/services/update/{id}": {
            "put": {
                "tags": [
                    "Service Routes"
                ],
                "summary": "Update Service",
                "operationId": "4986dc741681a39ffb0d9859cec73391",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "object id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "specialtyID": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Name of the new Service"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 34224
                                    },
                                    "shortDescription": {
                                        "type": "string",
                                        "example": "content of the new Service"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "author of the new Service"
                                    },
                                    "detail": {
                                        "type": "string",
                                        "example": "detail"
                                    },
                                    "discountPrice": {
                                        "type": "number",
                                        "example": 312412
                                    },
                                    "duration": {
                                        "type": "number",
                                        "example": 2342
                                    },
                                    "isHidden": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/services/delete/{id}": {
            "delete": {
                "tags": [
                    "Service Routes"
                ],
                "summary": "Update Service",
                "operationId": "a9a585b2a6e7e3d7b1015a0f679f9f44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "object id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
...
    },
...
}